# bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ZeraAPI, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/WishList/{id}": {
      "get": {
        "tags": [
          "WishList"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WishList"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WishList/create": {
      "post": {
        "tags": [
          "WishList"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/WhishList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WishList/update": {
      "put": {
        "tags": [
          "WishList"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/WhishList"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/WishListitem/{id}": {
      "get": {
        "tags": [
          "WishListitem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "WishListitem"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WishListitem/create": {
      "post": {
        "tags": [
          "WishListitem"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/WishListItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/WishListitem/update": {
      "put": {
        "tags": [
          "WishListitem"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/WishListItem"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "definitions": {
...
    "Cart": {
      "type": "object",
      "properties": {
        "cartId": {
          "type": "string"
        },
        "id": {
          "type": "string"
        },
        "cartItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartItem"
          }
        }
      },
      "additionalProperties": false
    },
...
    "CartItem": {
      "type": "object",
      "properties": {
        "cartItemId": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "cartId": {
          "type": "string"
        },
        "cart": {
          "$ref": "#/definitions/Cart"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "type": "string"
        },
        "product": {
          "$ref": "#/definitions/Product"
        }
      },
      "additionalProperties": false
    },
...
    "Order": {
      "required": [
        "orderDate",
        "shippingAddress",
        "shippingDate",
        "totalAmount"
      ],
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "orderDate": {
          "format": "date-time",
          "type": "string"
        },
        "shippingDate": {
          "format": "date-time",
          "type": "string"
        },
        "shippingAddress": {
          "type": "string"
        },
        "totalAmount": {
          "format": "double",
          "type": "number"
        },
        "status": {
          "$ref": "#/definitions/OrderStatus"
        },
        "paymentMethodId": {
          "format": "int32",
          "type": "integer"
        },
        "paymentMethod": {
          "$ref": "#/definitions/PaymentMethod"
        },
        "id": {
          "type": "string"
        },
        "userCustomised": {
          "$ref": "#/definitions/UserCustomised"
        }
      },
      "additionalProperties": false
    },
    "OrderItem": {
      "type": "object",
      "properties": {
        "orderItemId": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "$ref": "#/definitions/Order"
        }
      },
      "additionalProperties": false
    },
    "OrderStatus": {
      "format": "int32",
      "enum": [
        0,
        1,
        2,
        3
      ],
      "type": "integer"
    },
    "PaymentMethod": {
      "required": [
        "cardNumber",
        "cvv",
        "expiryDate",
        "method",
        "nameOnCard",
        "payPalEmail"
      ],
      "type": "object",
      "properties": {
        "paymentMethodId": {
          "format": "int32",
          "type": "integer"
        },
        "method": {
          "type": "string"
        },
        "cardNumber": {
          "type": "string"
        },
        "expiryDate": {
          "type": "string"
        },
        "cvv": {
          "type": "string"
        },
        "nameOnCard": {
          "type": "string"
        },
        "payPalEmail": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Product": {
      "required": [
        "description",
        "imageURL",
        "price",
        "productName",
        "producttype"
      ],
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productName": {
          "type": "string"
        },
        "producttype": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "description": {
          "type": "string"
        },
        "imageURL": {
          "type": "string"
        },
        "cartItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/CartItem"
          }
        },
        "orderItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderItem"
          }
        },
        "wishListItem": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WishListItem"
          }
        }
      },
      "additionalProperties": false
    },
...
    "UserCustomised": {
      "required": [
        "name"
      ],
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "userName": {
          "type": "string"
        },
        "normalizedUserName": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "normalizedEmail": {
          "type": "string"
        },
        "emailConfirmed": {
          "type": "boolean"
        },
        "passwordHash": {
          "type": "string"
        },
        "securityStamp": {
          "type": "string"
        },
        "concurrencyStamp": {
          "type": "string"
        },
        "phoneNumber": {
          "type": "string"
        },
        "phoneNumberConfirmed": {
          "type": "boolean"
        },
        "twoFactorEnabled": {
          "type": "boolean"
        },
        "lockoutEnd": {
          "format": "date-time",
          "type": "string"
        },
        "lockoutEnabled": {
          "type": "boolean"
        },
        "accessFailedCount": {
          "format": "int32",
          "type": "integer"
        },
        "name": {
          "maxLength": 100,
          "minLength": 0,
          "type": "string"
        },
        "address": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "WhishList": {
      "type": "object",
      "properties": {
        "whishListId": {
          "format": "int32",
          "type": "integer"
        },
        "id": {
          "type": "string"
        },
        "userCustomised": {
          "$ref": "#/definitions/UserCustomised"
        },
        "createdAt": {
          "format": "date-time",
          "type": "string"
        },
        "updatedAt": {
          "format": "date-time",
          "type": "string"
        },
        "wishListItems": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/WishListItem"
          }
        }
      },
      "additionalProperties": false
    },
    "WishListItem": {
      "type": "object",
      "properties": {
        "wishListItemId": {
          "format": "int32",
          "type": "integer"
        },
        "wishList": {
          "$ref": "#/definitions/WhishList"
        },
        "whishListId": {
          "format": "int32",
          "type": "integer"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "product": {
          "$ref": "#/definitions/Product"
        }
      },
      "additionalProperties": false
    }
  }
}